requirements.txt
Django==2.0.1
django-appconf==1.0.2
django-filter==1.1.0
django-imagekit==4.0.2
django-webpack-loader==0.5.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
pilkit==2.0
Pillow==5.0.0
PyJWT==1.5.3
pytz==2017.3
six==1.11.0
